refactor: organize server/ into functional sub-packages
Move 27 flat Python files into 7 sub-packages:
- core/ (config, models, state)
- chat/ (routes, streaming)
- runs/ (routes, helpers, sweep_routes, log_routes, evo_sweep)
- agent/ (wild_loop_v2, wild_routes, v2_prompts)
- skills/ (manager, routes)
- memory/ (store, routes)
- integrations/ (slack, git, journey, cluster, plan)
- tools/ (gpuwrap_detect, job_sidecar, mcp)

Update all import statements across server/, tests/, and build script.
Fix _SERVER_FILE_DIR to account for config.py moving one level deeper.
Fix wild_loop_v2.py import fallbacks for new package paths.
Add PyInstaller hidden imports for all new sub-packages.

All 96 tests pass.
