+++ server/agent/__init__.py
@@ -1,0 +1,1 @@
+# Wild loop / autonomous agent engine
