+++ server/core/__init__.py
@@ -1,0 +1,1 @@
+# Core shared modules: config, state, models
