+++ server/integrations/__init__.py
@@ -1,0 +1,1 @@
+# External integrations: slack, git, cluster, plan, journey
