+++ server/memory/__init__.py
@@ -1,0 +1,1 @@
+# Memory bank (persistent lessons / context)
