+++ server/skills/__init__.py
@@ -1,0 +1,1 @@
+# Prompt skill management
