+++ server/tools/__init__.py
@@ -1,0 +1,1 @@
+# Standalone CLI tools: sidecar, MCP, gpuwrap
