+++ server/tools/research_agent_mcp.py
