+++ server/tools/slurm_mcp_cli.py
